requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.27.0
google-generativeai==0.8.3
pypdf2==3.0.1
beautifulsoup4==4.12.3
python-multipart==0.0.12
pydantic-settings==2.5.2
pydantic[email]==2.9.2
